Refactor `TransitionPushedViewController` to extract some reusable logic out.
